fix: long not supported on return values

+++ backend-services/voting-ledger-follower-app/src/main/java/org/cardano/foundation/voting/repository/ProposalRepository.java
@@ -15,6 +15,6 @@ public interface ProposalRepository extends JpaRepository<Proposal, String> {
 
     @Query("DELETE FROM Proposal p WHERE p.absoluteSlot > :slot")
     @Modifying
-    long deleteAllAfterSlot(@Param("slot") long slot);
+    int deleteAllAfterSlot(@Param("slot") long slot);
 
 }
